Move JsonPatchDocument into the root util package
