+++ sdk/core/azure-core/src/main/java/com/azure/core/util/JsonPatchDocument.java
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.util.jsonpatch;
+package com.azure.core.util;
 
-import com.azure.core.util.Option;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.serializer.JacksonAdapter;
 import com.azure.core.util.serializer.JsonSerializer;
@@ -68,7 +67,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendAdd#String-Object}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendAdd#String-Object}
      *
      * @param path The path to apply the addition.
      * @param value The value that will be serialized and added to the path.
@@ -89,7 +88,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendAddRaw#String-String}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendAddRaw#String-String}
      *
      * @param path The path to apply the addition.
      * @param rawJson The raw JSON value that will be added to the path.
@@ -113,7 +112,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendReplace#String-Object}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendReplace#String-Object}
      *
      * @param path The path to replace.
      * @param value The value will be serialized and used as the replacement.
@@ -131,7 +130,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendReplaceRaw#String-String}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendReplaceRaw#String-String}
      *
      * @param path The path to replace.
      * @param rawJson The raw JSON value that will be used as the replacement.
@@ -155,7 +154,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendCopy#String-String}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendCopy#String-String}
      *
      * @param from The path to copy from.
      * @param path The path to copy to.
@@ -178,7 +177,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendMove#String-String}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendMove#String-String}
      *
      * @param from The path to move from.
      * @param path The path to move to.
@@ -199,7 +198,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendRemove#String}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendRemove#String}
      *
      * @param path The path to remove.
      * @return The updated JsonPatchDocument object.
@@ -218,7 +217,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendTest#String-Object}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendTest#String-Object}
      *
      * @param path The path to test.
      * @param value The value that will be serialized and used to test against.
@@ -236,7 +235,7 @@ public class JsonPatchDocument {
      *
      * <p><strong>Code Samples</strong></p>
      *
-     * {@codesnippet com.azure.core.util.jsonpatch.JsonPatchDocument.appendTestRaw#String-String}
+     * {@codesnippet com.azure.core.util.JsonPatchDocument.appendTestRaw#String-String}
      *
      * @param path The path to test.
      * @param rawJson The raw JSON value that will be used to test against.
