+++ sdk/core/azure-core/src/main/java/com/azure/core/util/JsonPatchDocumentSerializer.java
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.util.jsonpatch;
+package com.azure.core.util;
 
-import com.azure.core.util.CoreUtils;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
