--- sdk/core/azure-core/src/main/java/com/azure/core/util/jsonpatch/package-info.java
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-/**
- * Package containing JSON Patch classes.
- */
-package com.azure.core.util.jsonpatch;
