+++ sdk/core/azure-core/src/test/java/com/azure/core/util/JsonPatchDocumentTests.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.util.jsonpatch;
+package com.azure.core.util;
 
 import com.azure.core.util.serializer.JacksonAdapter;
 import com.azure.core.util.serializer.JsonSerializer;
